Resolving conflict with Candlestick
