+++ Candlestick/.eslintrc.json
