+++ Candlestick/index.js
@@ -3,11 +3,12 @@
  */
 
 import * as d3 from 'd3';
+import * as gLegend from 'g-legend';
 import gChartframe from 'g-chartframe';
 import * as gAxis from 'g-axis';
-import gChartcolour from 'g-chartcolour';
 import * as parseData from './parseData.js';
 import * as candlestick from './candlestick.js';
+import gChartcolour from 'g-chartcolour';
 
 const dataFile = 'data.csv';
 
@@ -39,76 +40,91 @@ const yAxisAlign = 'right';// alignment of the axis
 const xAxisAlign = 'bottom';// alignment of the axis
 const interval = 'days';// date interval on xAxis "century", "jubilee", "decade", "lustrum", "years","months","days"
 const annotate = true; /* show annotations, defined in the 'annotate' column */ // eslint-disable-line
-const markers = false; /* show dots on lines */ // eslint-disable-line
-const legendAlign = 'vert'; /* hori or vert, alignment of the legend */ // eslint-disable-line
-const legendType = 'line'; /* rect, line or circ, geometry of legend marker */ // eslint-disable-line
+const markers = false;/* show dots on lines */ // eslint-disable-line
+const legendAlign = 'vert';/* hori or vert, alignment of the legend */ // eslint-disable-line
+const legendType = 'line';/* rect, line or circ, geometry of legend marker */ // eslint-disable-line
 const minorAxis = false;// turns on or off the minor axis
 const highlightNames = []; // create an array names you want to highlight eg. ['series1','series2']
 const invertScale = false;
 const logScale = false; // eslint-disable-line
 const intraday = true;
 const chartColour = d3.scaleOrdinal()
-  .domain(Object.keys(gChartcolour.categorical_bar))
-  .range(Object.values(gChartcolour.categorical_bar));
+    .domain(Object.keys(gChartcolour.categorical_line))
+    .range(Object.values(gChartcolour.categorical_line));
 
 
 // Individual frame configuration, used to set margins (defaults shown below) etc
 const frame = {
     webS: gChartframe.webFrameS(sharedConfig)
- .margin({ top: 100, left: 15, bottom: 82, right: 5 })
- // .title('Put headline here') // use this if you need to override the defaults
- // .subtitle("Put headline |here") //use this if you need to override the defaults
- .height(400),
+        .margin({
+            top: 100, left: 15, bottom: 82, right: 5,
+        })
+    // .title('Put headline here') // use this if you need to override the defaults
+    // .subtitle("Put headline |here") //use this if you need to override the defaults
+        .height(400),
 
     webM: gChartframe.webFrameM(sharedConfig)
- .margin({ top: 100, left: 20, bottom: 86, right: 5 })
- // .title("Put headline here")
- .height(500),
+        .margin({
+            top: 100, left: 20, bottom: 86, right: 5,
+        })
+    // .title("Put headline here")
+        .height(500),
 
     webL: gChartframe.webFrameL(sharedConfig)
- .margin({ top: 100, left: 20, bottom: 104, right: 5 })
- // .title("Put headline here")
- .height(700)
- .fullYear(true),
+        .margin({
+            top: 100, left: 20, bottom: 104, right: 5,
+        })
+    // .title("Put headline here")
+        .height(700)
+        .fullYear(true),
 
     webMDefault: gChartframe.webFrameMDefault(sharedConfig)
- .margin({ top: 100, left: 20, bottom: 86, right: 5 })
- // .title("Put headline here")
- .height(500),
+        .margin({
+            top: 100, left: 20, bottom: 86, right: 5,
+        })
+    // .title("Put headline here")
+        .height(500),
 
     print: gChartframe.printFrame(sharedConfig)
- .margin({ top: 40, left: 7, bottom: 35, right: 7 })
-  // .title("Put headline here")
-  // .width(53.71)// 1 col
-  .width(112.25)// 2 col
-  // .width(170.8)// 3 col
-  // .width(229.34)// 4 col
-  // .width(287.88)// 5 col
-  // .width(346.43)// 6 col
-  // .width(74)// markets std print
-  .height(58.21), // markets std print
+        .margin({
+            top: 40, left: 7, bottom: 35, right: 7,
+        })
+    // .title("Put headline here")
+    // .width(53.71)// 1 col
+        .width(112.25)// 2 col
+    // .width(170.8)// 3 col
+    // .width(229.34)// 4 col
+    // .width(287.88)// 5 col
+    // .width(346.43)// 6 col
+    // .width(74)// markets std print
+        .height(58.21), // markets std print
 
     social: gChartframe.socialFrame(sharedConfig)
- .margin({ top: 140, left: 50, bottom: 138, right: 40 })
- // .title("Put headline here")
- .width(612)
- .height(612), // 700 is ideal height for Instagram
+        .margin({
+            top: 140, left: 50, bottom: 138, right: 40,
+        })
+    // .title("Put headline here")
+        .width(612)
+        .height(612), // 700 is ideal height for Instagram
 
     video: gChartframe.videoFrame(sharedConfig)
- .margin({ left: 207, right: 207, bottom: 210, top: 233 }),
- // .title("Put headline here")
+        .margin({
+            left: 207, right: 207, bottom: 210, top: 233,
+        }),
+    // .title("Put headline here")
 };
 
 
 // add the frames to the page...
 d3.selectAll('.framed')
-  .each(function addFrames() {
-      const figure = d3.select(this);
-      figure.select('svg')
-          .call(frame[figure.node().dataset.frame]);
-  });
-parseData.fromCSV(dataFile, dateStructure, { yMin, highlightNames })
-.then(({ plotData, valueExtent, highlights, annos }) => {
+    .each(function addFrames() {
+        const figure = d3.select(this);
+        figure.select('svg')
+            .call(frame[figure.node().dataset.frame]);
+    });
+parseData.fromCSV(dataFile, dateStructure, { yMin, highlightNames }).then(({
+    plotData, valueExtent, highlights, annos,
+}) => {
     Object.keys(frame).forEach((frameName) => {
         const currentFrame = frame[frameName];
 
@@ -122,8 +138,8 @@ parseData.fromCSV(dataFile, dateStructure, { yMin, highlightNames })
         const myAnnotations = candlestick.drawAnnotations();// sets up annotations
 
 
-          // .seriesNames(seriesNames)
-          // .highlightNames(highlightNames)
+        // .seriesNames(seriesNames)
+        // .highlightNames(highlightNames)
 
         // create a 'g' element at the back of the chart to add time period
         // highlights after axis have been created
@@ -133,19 +149,19 @@ parseData.fromCSV(dataFile, dateStructure, { yMin, highlightNames })
         const plotAnnotation = currentFrame.plot().append('g').attr('class', 'annotations-holder');
 
         myYAxis
-          .domain([Math.min(yMin, valueExtent[0]), Math.max(yMax, valueExtent[1])])
-          .range([currentFrame.dimension().height, 0])
-          .numTicks(numTicksy)
-          .tickSize(tickSize)
-          .yAxisHighlight(yAxisHighlight)
-          .align(yAxisAlign)
-          .frameName(frameName)
-          .invert(invertScale);
+            .domain([Math.min(yMin, valueExtent[0]), Math.max(yMax, valueExtent[1])])
+            .range([currentFrame.dimension().height, 0])
+            .numTicks(numTicksy)
+            .tickSize(tickSize)
+            .yAxisHighlight(yAxisHighlight)
+            .align(yAxisAlign)
+            .frameName(frameName)
+            .invert(invertScale);
 
         // Draw the yAxis first, this will position the yAxis correctly and
         // measure the width of the label text
         currentFrame.plot()
-          .call(myYAxis);
+            .call(myYAxis);
 
         // return the value in the variable newMargin
         if (yAxisAlign === 'right') {
@@ -176,21 +192,21 @@ parseData.fromCSV(dataFile, dateStructure, { yMin, highlightNames })
 
         // Set up xAxis for this frame
         myXAxis
-          .domain(xDomain)
-          .range([0, (currentFrame.dimension().width - (boxWidth / 2))])
-          .align(xAxisAlign)
-          .fullYear(false)
-          .interval(interval)
-          .tickSize(currentFrame.rem() * 0.75)
-          .minorAxis(minorAxis)
-          .minorTickSize(currentFrame.rem() * 0.3)
-          .fullYear(false)
-          .frameName(frameName)
-          .intraday(intraday);
+            .domain(xDomain)
+            .range([0, (currentFrame.dimension().width - (boxWidth / 2))])
+            .align(xAxisAlign)
+            .fullYear(false)
+            .interval(interval)
+            .tickSize(currentFrame.rem() * 0.75)
+            .minorAxis(minorAxis)
+            .minorTickSize(currentFrame.rem() * 0.3)
+            .fullYear(false)
+            .frameName(frameName)
+            .intraday(intraday);
 
         // Draw the xAxis
         currentFrame.plot()
-          .call(myXAxis);
+            .call(myXAxis);
 
         if (xAxisAlign === 'bottom') {
             myXAxis.xLabel().attr('transform', `translate(0,${currentFrame.dimension().height})`);
@@ -203,49 +219,49 @@ parseData.fromCSV(dataFile, dateStructure, { yMin, highlightNames })
         }
 
         myChart
-          .yScale(myYAxis.scale())
-          .xScale(myXAxis.scale())
-          .plotDim(currentFrame.dimension())
-          .rem(currentFrame.rem())
-          .colourPalette(chartColour)
-          .intraday(intraday)
-          .boxWidth(boxWidth);
+            .yScale(myYAxis.scale())
+            .xScale(myXAxis.scale())
+            .plotDim(currentFrame.dimension())
+            .rem(currentFrame.rem())
+            .colourPalette(chartColour)
+            .intraday(intraday)
+            .boxWidth(boxWidth);
 
         currentFrame.plot()
-          .selectAll('lines')
-          .data(plotData)
-          .enter()
-          .append('g')
-          .attr('class', 'whisker')
-          .attr('id', d => d.name)
-          .call(myChart);
+            .selectAll('.candlesticks')
+            .data(plotData)
+            .enter()
+            .append('g')
+            .attr('class', 'candlesticks')
+            .attr('id', d => d.name)
+            .call(myChart);
 
         // Set up highlights for this frame
         myHighlights
-          .yScale(myYAxis.scale())
-          .xScale(myXAxis.scale());
+            .yScale(myYAxis.scale())
+            .xScale(myXAxis.scale());
 
         // Draw the highlights before the lines and xAxis
         axisHighlight
-          .selectAll('.highlights')
-          .data(highlights)
-          .enter()
-          .append('g')
-          .call(myHighlights);
+            .selectAll('.highlights')
+            .data(highlights)
+            .enter()
+            .append('g')
+            .call(myHighlights);
 
         // Set up highlights for this frame
         myAnnotations
-          .yScale(myYAxis.scale())
-          .xScale(myXAxis.scale())
-          .rem(currentFrame.rem());
+            .yScale(myYAxis.scale())
+            .xScale(myXAxis.scale())
+            .rem(currentFrame.rem());
 
         // Draw the annotations before the lines
         plotAnnotation
-          .selectAll('.annotation')
-          .data(annos)
-          .enter()
-          .append('g')
-          .call(myAnnotations);
+            .selectAll('.annotation')
+            .data(annos)
+            .enter()
+            .append('g')
+            .call(myAnnotations);
     });
     // addSVGSavers('figure.saveable');
 });
